refactor: 리스트 다중 삭제 API 위치 변경 및 코드 리펙토링 (#193)
