+++ src/main/java/com/listywave/list/presentation/dto/request/ListsDeleteRequest.java
@@ -1,0 +1,8 @@
+package com.listywave.list.presentation.dto.request;
+
+import java.util.List;
+
+public record ListsDeleteRequest(
+        List<Long> listId
+) {
+}
