+++ src/main/java/com/listywave/user/presentation/UserController.java
@@ -10,7 +10,6 @@ import com.listywave.user.application.dto.RecommendUsersResponse;
 import com.listywave.user.application.dto.UserInfoResponse;
 import com.listywave.user.application.dto.search.AllUserSearchResponse;
 import com.listywave.user.application.service.UserService;
-import com.listywave.user.presentation.dto.UserListsDeleteRequest;
 import com.listywave.user.presentation.dto.UserProfileUpdateRequest;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -141,14 +140,5 @@ public class UserController {
     ResponseEntity<Void> withdraw(@Auth Long userId) {
         userService.withdraw(userId);
         return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/users/lists")
-    ResponseEntity<Void> deleteLists(
-            @Auth Long userId,
-            @RequestBody UserListsDeleteRequest request
-    ) {
-        userService.deleteLists(userId, request.listId());
-        return ResponseEntity.ok().build();
     }
 }
