--- src/main/java/com/listywave/user/presentation/dto/UserListsDeleteRequest.java
@@ -1,8 +1,0 @@
-package com.listywave.user.presentation.dto;
-
-import java.util.List;
-
-public record UserListsDeleteRequest(
-        List<Long> listId
-) {
-}
